refactor(gfn): modularize SDP transformations
