+++ opennow-stable/src/renderer/src/platforms/gfn/sdp.test.ts
@@ -3,261 +3,20 @@
 import test from "node:test";
 import assert from "node:assert/strict";
 
-import {
-  buildNvstSdp,
-  fixServerIp,
-  mungeAnswerSdp,
-  preferCodec,
-  rewriteIceCandidateEndpoint,
-  rewriteH265LevelIdByProfile,
-  rewriteH265TierFlag,
-  rewriteSdpIceCandidateEndpoints,
-} from "./sdp";
+import * as sdp from "./sdp";
 
-test("fixServerIp replaces 0.0.0.0 candidate IPs without changing connection lines", () => {
-  const sdp = [
-    "v=0",
-    "c=IN IP4 0.0.0.0",
-    "a=candidate:1 1 udp 2130706431 0.0.0.0 47998 typ host",
-    "a=candidate:2 1 tcp 1 192.168.1.5 9 typ host",
-  ].join("\n");
-
-  const fixed = fixServerIp(sdp, "161-248-11-132.bpc.geforcenow.nvidiagrid.net");
-
-  assert.match(fixed, /c=IN IP4 0\.0\.0\.0/);
-  assert.match(fixed, /a=candidate:1 1 udp 2130706431 161\.248\.11\.132 47998 typ host/);
-  assert.match(fixed, /a=candidate:2 1 tcp 1 192\.168\.1\.5 9 typ host/);
-  assert.equal(fixServerIp(sdp, "unparseable.example.com"), sdp);
-});
-
-test("rewriteSdpIceCandidateEndpoints points server candidates at WebRTC mediaConnectionInfo", () => {
-  const sdp = [
-    "v=0",
-    "c=IN IP4 0.0.0.0",
-    "a=candidate:1 1 udp 2122260223 203.0.113.10 47998 typ host",
-    "a=candidate:2 1 tcp 1518214911 203.0.113.10 9 typ host tcptype active",
-  ].join("\r\n");
-
-  const rewritten = rewriteSdpIceCandidateEndpoints(sdp, {
-    ip: "198.51.100.55",
-    port: 18784,
-    usage: 2,
-  });
-
-  assert.equal(rewritten.replacements, 2);
-  assert.match(rewritten.sdp, /a=candidate:1 1 udp 2122260223 198\.51\.100\.55 18784 typ host/);
-  assert.match(rewritten.sdp, /a=candidate:2 1 tcp 1518214911 198\.51\.100\.55 18784 typ host tcptype active/);
-  assert.match(rewritten.sdp, /c=IN IP4 0\.0\.0\.0/);
-  assert.match(rewritten.sdp, /\r\n/);
-});
-
-test("rewriteIceCandidateEndpoint skips non-WebRTC mediaConnectionInfo usages", () => {
-  const candidate = "candidate:1 1 udp 2122260223 203.0.113.10 47998 typ host";
-
-  assert.deepEqual(
-    rewriteIceCandidateEndpoint(candidate, { ip: "198.51.100.55", port: 18784, usage: 14 }),
-    { candidate, rewritten: false },
-  );
-  assert.deepEqual(
-    rewriteIceCandidateEndpoint(candidate, { ip: "198.51.100.55", port: 18784, usage: 17 }),
-    {
-      candidate: "candidate:1 1 udp 2122260223 198.51.100.55 18784 typ host",
-      rewritten: true,
-    },
-  );
-});
-
-test("preferCodec keeps selected video payloads and RTX apt payloads while leaving audio untouched", () => {
-  const sdp = [
-    "v=0",
-    "m=audio 9 UDP/TLS/RTP/SAVPF 111 0",
-    "a=rtpmap:111 opus/48000/2",
-    "a=fmtp:111 minptime=10;useinbandfec=1",
-    "m=video 9 UDP/TLS/RTP/SAVPF 96 97 98 99 100 101 102",
-    "a=rtpmap:96 H264/90000",
-    "a=rtcp-fb:96 nack",
-    "a=rtpmap:97 rtx/90000",
-    "a=fmtp:97 apt=96",
-    "a=rtpmap:98 H265/90000",
-    "a=fmtp:98 profile-id=1;tier-flag=0;level-id=153",
-    "a=rtcp-fb:98 nack pli",
-    "a=rtpmap:99 rtx/90000",
-    "a=fmtp:99 apt=98",
-    "a=rtpmap:100 AV1/90000",
-    "a=rtpmap:101 flexfec-03/90000",
-    "a=rtpmap:102 ulpfec/90000",
-  ].join("\n");
-
-  const filtered = preferCodec(sdp, "H265");
-
-  assert.match(filtered, /m=video 9 UDP\/TLS\/RTP\/SAVPF 98 99/);
-  assert.match(filtered, /a=rtpmap:98 H265\/90000/);
-  assert.match(filtered, /a=rtpmap:99 rtx\/90000/);
-  assert.match(filtered, /a=fmtp:99 apt=98/);
-  assert.doesNotMatch(filtered, /a=rtpmap:96 H264/);
-  assert.doesNotMatch(filtered, /a=rtpmap:100 AV1/);
-  assert.doesNotMatch(filtered, /flexfec|ulpfec/);
-  assert.match(filtered, /m=audio 9 UDP\/TLS\/RTP\/SAVPF 111 0/);
-  assert.match(filtered, /a=fmtp:111 minptime=10;useinbandfec=1/);
-});
-
-test("rewrites H265 tier flag and clamps oversized level by profile", () => {
-  const sdp = [
-    "m=video 9 UDP/TLS/RTP/SAVPF 98 99",
-    "a=rtpmap:98 H265/90000",
-    "a=fmtp:98 profile-id=1;tier-flag=1;level-id=186",
-    "a=rtpmap:99 H265/90000",
-    "a=fmtp:99 profile-id=2;tier-flag=1;level-id=255",
-  ].join("\n");
-
-  const tier = rewriteH265TierFlag(sdp, 0);
-  const level = rewriteH265LevelIdByProfile(tier.sdp, { 1: 153, 2: 186 });
-
-  assert.equal(tier.replacements, 2);
-  assert.equal(level.replacements, 2);
-  assert.match(level.sdp, /a=fmtp:98 profile-id=1;tier-flag=0;level-id=153/);
-  assert.match(level.sdp, /a=fmtp:99 profile-id=2;tier-flag=0;level-id=186/);
-});
-
-test("mungeAnswerSdp injects bitrate lines and appends opus stereo once", () => {
-  const sdp = [
-    "m=video 9 UDP/TLS/RTP/SAVPF 98",
-    "c=IN IP4 127.0.0.1",
-    "m=audio 9 UDP/TLS/RTP/SAVPF 111",
-    "a=fmtp:111 minptime=10;useinbandfec=1",
-  ].join("\n");
-
-  const munged = mungeAnswerSdp(sdp, 50000);
-  assert.match(munged, /m=video.*\nb=AS:50000\n/);
-  assert.match(munged, /m=audio.*\nb=AS:128\n/);
-  assert.match(munged, /a=fmtp:111 minptime=10;useinbandfec=1;stereo=1/);
-
-  const alreadyStereo = mungeAnswerSdp("m=audio 9 UDP/TLS/RTP/SAVPF 111\nb=AS:128\na=fmtp:111 minptime=10;stereo=1", 50000);
-  assert.equal((alreadyStereo.match(/stereo=1/g) ?? []).length, 1);
-  assert.equal((alreadyStereo.match(/b=AS:128/g) ?? []).length, 1);
-});
-
-test("buildNvstSdp includes stream quality and partially reliable input parameters", () => {
-  const sdp = buildNvstSdp({
-    width: 2560,
-    height: 1440,
-    fps: 120,
-    maxBitrateKbps: 80000,
-    partialReliableThresholdMs: 16,
-    codec: "AV1",
-    colorQuality: "10bit_444",
-    credentials: {
-      ufrag: "ufrag-test",
-      pwd: "password-test",
-      fingerprint: "AA:BB:CC",
-    },
-    hidDeviceMask: 128,
-    enablePartiallyReliableTransferGamepad: 15,
-    enablePartiallyReliableTransferHid: 128,
-  });
-
-  for (const line of [
-    "a=video.clientViewportWd:2560",
-    "a=video.clientViewportHt:1440",
-    "a=video.maxFPS:120",
-    "a=video.initialBitrateKbps:20000",
-    "a=video.initialPeakBitrateKbps:20000",
-    "a=vqos.bw.maximumBitrateKbps:80000",
-    "a=vqos.bw.minimumBitrateKbps:4000",
-    "a=vqos.bw.peakBitrateKbps:80000",
-    "a=vqos.drc.minRequiredBitrateCheckEnabled:1",
-    "a=vqos.fec.repairPercent:5",
-    "a=vqos.fec.repairMaxPercent:35",
-    "a=packetPacing.enableAccurateSleep:1",
-    "a=packetPacing.minNumPacketsPerGroup:15",
-    "a=video.framePacing.mode:2",
-    "a=video.framePacing.pid.minTargetFrameTimeUs:7916",
-    "a=video.adaptiveQuantization.spatialAQSetting:7",
-    "a=video.rtpNackQueueLength:1024",
-    "a=video.rtpNackQueueMaxPackets:512",
-    "a=video.rtpNackMaxPacketCount:25",
-    "a=aqos.enableRedundancy:1",
-    "a=aqos.redundancyLevel:2",
-    "a=ri.timestampsEnabled:1",
-    "a=video.bitDepth:10",
-    "a=vqos.drc.enable:0",
-    "a=vqos.dfc.enable:1",
-    "a=vqos.dfc.qpMaxResThresholdAdj:20",
-    "a=vqos.grc.qpMaxResThresholdAdj:20",
-    "a=vqos.resControl.cpmRtc.enable:0",
-    "a=ri.partialReliableThresholdMs:16",
-    "a=ri.hidDeviceMask:128",
-    "a=ri.enablePartiallyReliableTransferGamepad:15",
-    "a=ri.enablePartiallyReliableTransferHid:128",
-  ]) {
-    assert.match(sdp, new RegExp(line.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")));
-  }
-});
-
-test("buildNvstSdp applies the official 90 FPS capture profile", () => {
-  const sdp = buildNvstSdp({
-    width: 1280,
-    height: 800,
-    fps: 90,
-    maxBitrateKbps: 50000,
-    partialReliableThresholdMs: 16,
-    codec: "H265",
-    colorQuality: "10bit_420",
-    credentials: {
-      ufrag: "ufrag-test",
-      pwd: "password-test",
-      fingerprint: "AA:BB:CC",
-    },
-  });
-
-  const lines = new Set(sdp.split(/\r?\n/));
-  for (const line of [
-    "a=video.maxFPS:90",
-    "a=video.framePacing.pid.minTargetFrameTimeUs:10555",
-    "a=vqos.dfc.enable:1",
-    "a=vqos.dfc.dfcAlgoVersion:1",
-    "a=vqos.dfc.minTargetFps:60",
-    "a=video.fbcDynamicFpsGrabTimeoutMs:9",
-    "a=vqos.resControl.cpmRtc.decodeTimeThresholdMs:11",
-  ]) {
-    assert.equal(lines.has(line), true, `missing exact SDP line: ${line}`);
-  }
-  assert.equal(lines.has("a=vqos.maxStreamFpsEstimate:240"), false);
-});
-
-test("buildNvstSdp keeps dynamic split encode updates enabled for 240 FPS by default", () => {
-  const defaultSdp = buildNvstSdp({
-    width: 1920,
-    height: 1080,
-    fps: 240,
-    maxBitrateKbps: 75000,
-    partialReliableThresholdMs: 16,
-    codec: "H265",
-    colorQuality: "10bit_420",
-    credentials: {
-      ufrag: "ufrag-test",
-      pwd: "password-test",
-      fingerprint: "AA:BB:CC",
-    },
-  });
-  const diagnosticOffSdp = buildNvstSdp({
-    width: 1920,
-    height: 1080,
-    fps: 240,
-    maxBitrateKbps: 75000,
-    partialReliableThresholdMs: 16,
-    codec: "H265",
-    colorQuality: "10bit_420",
-    credentials: {
-      ufrag: "ufrag-test",
-      pwd: "password-test",
-      fingerprint: "AA:BB:CC",
-    },
-    dynamicSplitEncodeUpdatesEnabled: false,
-  });
-
-  assert.match(defaultSdp, /a=video\.updateSplitEncodeStateDynamically:1/);
-  assert.match(defaultSdp, /a=video\.videoSplitEncodeStripsPerFrame:63/);
-  assert.match(defaultSdp, /a=vqos\.relaxMaxBitrate\.iirFilterFactor:120/);
-  assert.match(diagnosticOffSdp, /a=video\.updateSplitEncodeStateDynamically:0/);
+test("sdp barrel preserves the public API", () => {
+  assert.deepEqual(Object.keys(sdp).sort(), [
+    "buildNvstSdp",
+    "extractIceCredentials",
+    "extractIceUfragFromOffer",
+    "extractPublicIp",
+    "fixServerIp",
+    "mungeAnswerSdp",
+    "preferCodec",
+    "rewriteH265LevelIdByProfile",
+    "rewriteH265TierFlag",
+    "rewriteIceCandidateEndpoint",
+    "rewriteSdpIceCandidateEndpoints",
+  ]);
 });
